// geekstack-app/src/app/core/model/card-dragonballzfw.model.ts
export interface CardDragonBallZFW {
    _id: string;
    awakenform: boolean;
    booster: string;
    cardId: string;
    cardUid: string;
    cardName: string;
    cardNameLower: string;
    cardtype: string;
    cardtypeLower: string;
    color: string;
    colorLower: string;
    combopower: string;
    cost: string;
    effects: string;
    features: string;
    featuresLower: string;
    image: string;
    power: string;
    rarity: string;
    raritySub: string;
    setFrom: string;
    specifiedcost: string;
    urlimage: string;
    YYT: string;
    price_yyt_id: string;
  }
  