// geekstack-app/src/app/core/model/card-cookierunbraverse.model.ts
export interface CookieRunCard {
    _id: string;

    id: number;

    elementId: number;

    title: string;

    artistTitle: string;

    productTitle: string;

    cardDescription: string;

    rarity: string;

    hp: string;

    cardNo: string;

    grade: string;

    urlimage: string;

    productCategory: string;

    productCategoryTitle: string;

    cardType: string;

    cardTypeTitle: string;

    energyType: string;

    energyTypeTitle: string;

    cardLevel: string;

    cardLevelTitle: string;

    boostercode: string;

    cardUid: String;

    cardId: String;

}